Require message acknowledgment in worker.rb
Message acknowledgments are turned off by default. It's time to turn
them on using the :manual_ack option and send a proper acknowledgment
from the worker, once we're done with a task.

Using this code we can be sure that even if you kill a worker using
CTRL+C while it was processing a message, nothing will be lost. Soon
after the worker dies all unacknowledged messages will be redelivered.