Make messages durable
Two things are required to make sure that messages aren't lost when
RabbitMQ quits or crashes: we need to mark both the queue and messages
as durable.First, we need to make sure that RabbitMQ will never lose our
queue. In order to do so, we need to declare it as durable. RabbitMQ
doesn't allow you to redefine an existing queue with different
parameters and will return an error to any program that tries to do
that. But there is a quick workaround - let's declare a queue with
different name, for example task_queue