Add prefetch to worker
To prevent some workers from being overloaded due to uneven message
complexities, we can use the prefetch method with the value of 1. This
tells RabbitMQ not to give more than one message to a worker at a time.
Or, in other words, don't dispatch a new message to a worker until it
has processed and acknowledged the previous one. Instead, it will
dispatch it to the next worker that is not still busy.